move stuff to hardware folder
