+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Blue_NetZone_RR_Auto.java
@@ -6,6 +6,8 @@ import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
+import org.firstinspires.ftc.teamcode.hardware.PinpointDrive;
+
 @Autonomous(name = "ODOMETRY: Robot Auto Drive 3 Sample Net Zone + LV 1 Ascent", group = "Linear OpMode")
 public final class Blue_NetZone_RR_Auto extends LinearOpMode {
     @Override
